fix(backend): keep task out of REVIEW when cancel escalates
When the lifecycle manager force-unblocks a hung agent, the synthetic
error bubbled up through waitForPromptDone into PromptTask's error
handler. isTransientPromptError doesn't match "cancel escalated", so
UpdateTaskState(REVIEW) fired — putting the task card in REVIEW even
though the user had just cancelled. Service.CancelAgent only reconciles
session state, not task state, so there was no recovery.

Wrap ErrCancelEscalated in waitForPromptDone's returned error so callers
can identify it via errors.Is, and exempt it from the REVIEW transition
in PromptTask.

Also splits the new cancel-escalation tests into a dedicated file so the
lifecycle test file stays under the 800-line revive limit (this was the
CI lint failure on the previous commit).
